Update WXApiObject.h
change uint32 to uint64

+++ ios/WXApiObject.h
@@ -128,7 +128,7 @@ typedef NS_ENUM(UInt64, enAppSupportContentFlag)
 /** 随机串，防重发 */
 @property (nonatomic, retain) NSString *nonceStr;
 /** 时间戳，防重发 */
-@property (nonatomic, assign) UInt32 timeStamp;
+@property (nonatomic, assign) UInt64 timeStamp;
 /** 商家根据财付通文档填写的数据和签名 */
 @property (nonatomic, retain) NSString *package;
 /** 商家根据微信开放平台文档对数据做的签名 */
